review: address Copilot feedback on cell/abilities tests
Two unaddressed comments from PR #178:

- damage_apply.rs grew to 724 lines after the inline test mod, just
  over the 700-line hard cap in CLAUDE.md. Convert to a directory
  module (damage_apply/mod.rs + damage_apply/tests.rs) — natural
  seam is "production code | tests", and the test mod was 309 lines.
  mod.rs is now 417 lines, tests.rs is 309, both under the cap.
- use_ability.rs out-of-range test matched method_index: 121 by
  literal. Switch to a guard-based pattern using
  crate::mercury::method_idx::ON_ERROR_CODE so the test doesn't
  drift if indices ever get reorganized, and so the intent is
  self-documenting.

All 38 cell::abilities tests still pass locally; clippy + fmt clean.
